Жень, у тебя в тестах был некомпилирующийся код - посмотри их, я чуть исправила явное привидения от int к short и пару методов у тебя тестируются, а в классе нет ZLColor - я их закомментировала
git-svn-id: https://only.mawhrin.net/repos/FBReaderJ/trunk@163 6a642e6f-84f6-412e-ac94-c4a38d5a04b0

+++ test/org/test/zlibrary/options/ModelTests.java
@@ -32,7 +32,7 @@ public class ModelTests extends TestCase{
 	private final double myDefaultDouble = 1.2;
     
 	public void setUp(){
-		myColorOption = new ZLColorOption("","","my Color", new ZLColor (176, 255, 0)); 
+		myColorOption = new ZLColorOption("","","my Color", new ZLColor ((short)176, (short)255, (short)0)); 
 		myDoubleOption = new ZLDoubleOption("","","My Double", myDefaultDouble); 
 		myIntegerOption = new ZLIntegerOption("","","I", myDefaultInt); 
 		myIntegerRangeOption = new ZLIntegerRangeOption("","","IR", -90L, 90L, myDefaultIntRange);
